Fixed sdist installation
